Loading Qtables & Command line arguments
Loading Q-tables is possible again. Some code got messed up.
The program asks user input when a table cannot be loaded.
Also, it is now possible to add the parameter file and the save option
as a command line argument
TCLI_log contains the used learning rate for reference.
The learning interface (and TorcsWorld) have a #define option to add -1
and 1 to the action space.